AIMonk-ObjectDetection-Microservices-assignment/ui-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
requests==2.32.4
pillow==10.4.0
python-multipart==0.0.20
